# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Hull Local API
  version: "1.0.0"
  description: |
    The contract between `hulld` and its clients (CLI, and the native GUIs from
    Phase 3 Wave 2 onward). This document is the FROZEN contract: every native
    client builds against it. Keep it in sync with `internal/api` , the
    coverage test `internal/api/openapi_test.go` fails if a registered route is
    missing here.

    Transport (ADR 0006): HTTP on `127.0.0.1:<port>` where `<port>` and a bearer
    token are read from `~/.hull/daemon.json` (written 0600 by a running
    daemon). Auth accepts `Authorization: Bearer <token>` or `?token=<token>`
    (the latter for EventSource/SSE, which cannot set headers). The server
    reflects any Origin for CORS , safe because the token, not the origin, is
    the access control and it only listens on loopback.

    Long operations (create, import, add-service, link, rebuild, reset) are
    JOBS: a POST returns a job reference; observe it by polling `GET
    /v1/jobs/{id}` or streaming `GET /v1/jobs/{id}/stream`.
servers:
  - url: http://127.0.0.1:{port}
    description: Local daemon (port from ~/.hull/daemon.json)
    variables:
      port:
        default: "0"
security:
  - bearerAuth: []
  - tokenQuery: []
...
paths:
...
  /v1/projects/{name}/unlink:
    post:
      tags: [services]
      summary: Remove a service link (db/redis/...) from a project
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UnlinkRequest" }
      responses:
        "204": { description: Unlinked }
...
  /v1/services:
    get:
      tags: [services]
      summary: List shared service instances
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ServiceInfo" }
    post:
      tags: [services]
      summary: Create and start a shared instance (job , image pull)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AddServiceRequest" }
      responses:
        "202":
          description: Job started
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobRef" }
  /v1/services/{name}/{action}:
    post:
      tags: [services]
      summary: Start, stop, or link a shared instance
      description: "`link` runs as a job; `start`/`stop` are synchronous."
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
        - name: action
          in: path
          required: true
          schema: { type: string, enum: [start, stop, link] }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LinkRequest" }
      responses:
        "204": { description: Done (start/stop) }
        "202":
          description: Job started (link)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobRef" }
  /v1/services/{name}:
    delete:
      tags: [services]
      summary: Destroy a shared instance and its data
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Destroyed }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    tokenQuery:
      type: apiKey
      in: query
      name: token
  schemas:
...
    UnlinkRequest:
      type: object
      required: [key]
      properties:
        key: { type: string }
    ServiceInfo:
      type: object
      required: [name, engine, version, container, running]
      properties:
        name: { type: string }
        engine: { type: string }
        version: { type: string }
        container: { type: string }
        running: { type: boolean }
        url: { type: string }
        host: { type: string }
        host_port: { type: integer }
        username: { type: string }
        linked_projects: { type: array, items: { type: string } }
    AddServiceRequest:
      type: object
      required: [engine]
      properties:
        engine: { type: string }
        version: { type: string }
    LinkRequest:
      type: object
      required: [project]
      properties:
        project: { type: string }
...
    Job:
      type: object
      required: [id, kind, status, lines, created]
      properties:
        id: { type: string }
        kind: { type: string }
        status: { type: string, enum: [running, done, failed] }
        error: { type: string }
        lines: { type: array, items: { type: string } }
        created: { type: string, format: date-time }
    JobRef:
      type: object
      required: [job]
      properties:
        job: { $ref: "#/components/schemas/Job" }
...